ecommerce-api/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.1
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
